Save refactor process
