+++ Assets/Player/PlayerInput.cs
@@ -1,0 +1,31 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PlayerInput : MonoBehaviour
+{
+    [HideInInspector] 
+    public float horizontalInput = 0;
+    [HideInInspector] 
+    public float verticalInput = 0;
+    [HideInInspector] 
+    public bool jumpPressed;
+    [HideInInspector] 
+    public bool grabPressed;
+    [HideInInspector] 
+    public bool climbUp;
+    [HideInInspector] 
+    public bool climbDown;
+
+    // Update is called once per frame
+    void Update()
+    {
+        horizontalInput = Input.GetAxisRaw("Horizontal");
+        verticalInput = Input.GetAxisRaw("Vertical");
+        
+        jumpPressed = Input.GetKeyDown(KeyCode.Space);
+        grabPressed = Input.GetKey(KeyCode.LeftControl);
+        climbUp = Input.GetKey(KeyCode.UpArrow);
+        climbDown = Input.GetKey(KeyCode.DownArrow);
+    }
+}
