+++ Assets/Player/PlayerJump.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: db29f0725b502e942b0df0390e846201
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
