+++ Assets/Player/PlayerParticles.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d66b778db8824c84587d5e9a65a008ab
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
